feat(backend): adapt backend to frienships management

+++ backend/src/users/users.service.ts
@@ -4,14 +4,20 @@ import { PageDto, PageOptionsDto } from '@dtos/pages';
 import { CreateUserDto, QueryUserDto, UpdateUserDto } from '@dtos/users';
 import { User } from './entities/user.entity';
 import * as fs from 'fs';
-import { ResponseFriendshipDto, FriendshipTypeEnum } from '@dtos/friendships';
-import { ResponseBlockDto, BlockTypeEnum } from '@dtos/blocks';
+import {
+  ResponseFriendshipDto,
+  FriendshipTypeEnum,
+  ListFriendshipDto,
+} from '@dtos/friendships';
+import { BlockTypeEnum, ListBlockDto } from '@dtos/blocks';
 import { AchievementsLogDto } from '@dtos/achievements-log';
 import { Block } from 'src/relationships/entities/block.entity';
 import { Friendship } from 'src/relationships/entities/friendship.entity';
 import { paginate } from 'src/common/paginate';
 import {
   DeleteResult,
+  FindManyOptions,
+  In,
   EntityNotFoundError,
   Repository,
   UpdateResult,
@@ -83,6 +89,9 @@ export class UsersService {
   }
 
   remove_avatar(filePath: string, reason: string) {
+    if (!fs.existsSync(filePath)) {
+      return;
+    }
     fs.unlink(filePath, (err) => {
       if (err) this.logger.error(err);
       else this.logger.log('Deleted ' + filePath + reason);
@@ -90,6 +99,7 @@ export class UsersService {
   }
 
   async updateAvatar(userId: number, filepath: string) {
+    // console.log('filepath update: ' + filepath);
     const user: User = await this.usersRepository.findOneByOrFail({
       id: userId,
     });
@@ -110,23 +120,55 @@ export class UsersService {
     return { fileStream: new StreamableFile(stream), ext: ext };
   }
 
-  findFriendships(id: number): Promise<ResponseFriendshipDto[]> {
-    return this.friendshipRepository.find({
-      where: [
-        { targetId: id, status: FriendshipTypeEnum.ACCEPTED },
-        { sourceId: id, status: FriendshipTypeEnum.ACCEPTED },
-      ],
+  async findFriendships(
+    id: number,
+    mode: number,
+  ): Promise<ListFriendshipDto[]> {
+    const options: FindManyOptions = mode
+      ? {
+          where: [
+            { targetId: id, status: FriendshipTypeEnum.ACCEPTED },
+            { sourceId: id, status: FriendshipTypeEnum.ACCEPTED },
+          ],
+        }
+      : {
+          where: { targetId: id, status: FriendshipTypeEnum.PENDING },
+        };
+
+    const tmp: ResponseFriendshipDto[] = await this.friendshipRepository.find(
+      options,
+    );
+
+    const ids: number[] = [];
+    const ret: ListFriendshipDto[] = [];
+
+    for (let i = 0; i < tmp.length; i++) {
+      ids.push(tmp[i].targetId == id ? tmp[i].sourceId : tmp[i].targetId);
+    }
+
+    const friends: User[] = await this.usersRepository.find({
+      where: {
+        id: In(ids),
+      },
     });
+
+    for (let i = 0; i < tmp.length; i++) {
+      for (let j = 0; j < friends.length; j++) {
+        if (ids[i] == friends[j].id) {
+          const uUd: UpdateUserDto = {
+            username: friends[j].username,
+            avatar: friends[j].avatar,
+          };
+          ret[i] = new ListFriendshipDto(tmp[i], uUd);
+          break;
+        }
+      }
+    }
+    return ret;
   }
 
-  findFriendRequests(id: number): Promise<ResponseFriendshipDto[]> {
-    return this.friendshipRepository.find({
-      where: { targetId: id, status: FriendshipTypeEnum.PENDING },
-    });
-  }
-
-  findBlocks(id: number): Promise<ResponseBlockDto[]> {
-    return this.blockRepository.find({
+  async findBlocks(id: number): Promise<ListBlockDto[]> {
+    const tmp: Block[] = await this.blockRepository.find({
       where: [
         { sourceId: id, status: BlockTypeEnum.MUTUAL },
         { targetId: id, status: BlockTypeEnum.MUTUAL },
@@ -134,6 +176,32 @@ export class UsersService {
         { targetId: id, status: BlockTypeEnum.T_BLOCKS_S },
       ],
     });
+    const ids: number[] = [];
+    const ret: ListBlockDto[] = [];
+
+    for (let i = 0; i < tmp.length; i++) {
+      ids.push(tmp[i].targetId == id ? tmp[i].sourceId : tmp[i].targetId);
+    }
+    const blocks: User[] = await this.usersRepository.find({
+      where: {
+        id: In(ids),
+      },
+    });
+
+    for (let i = 0; i < tmp.length; i++) {
+      for (let j = 0; j < blocks.length; j++) {
+        if (ids[i] == blocks[j].id) {
+          ret[i] = {
+            username: blocks[i].username,
+            avatar: blocks[i].avatar,
+            id: tmp[i].id,
+          };
+          blocks.splice(j);
+          break;
+        }
+      }
+    }
+    return ret;
   }
 
   findUnlockedAchievements(id: number): Promise<AchievementsLogDto[]> {
